+++ examples/msw/src/gen/msw/store/Handlers/placeOrderPatchHandler.ts
@@ -1,8 +1,9 @@
 import type { PlaceOrderPatchMutationResponse } from '../../../models/PlaceOrderPatch.ts'
 import { http } from 'msw'
 
-export function placeOrderPatchHandler(data?: PlaceOrderPatchMutationResponse) {
+export function placeOrderPatchHandler(data?: PlaceOrderPatchMutationResponse | ((info: Parameters<Parameters<typeof http.patch>[1]>[0]) => Response)) {
   return http.patch('*/store/order', function handler(info) {
+    if (typeof data === 'function') return data(info)
     return new Response(JSON.stringify(data), {
       headers: {
         'Content-Type': 'application/json',