+++ examples/msw/src/gen/msw/user/Handlers/getUserByNameHandler.ts
@@ -1,8 +1,9 @@
 import type { GetUserByNameQueryResponse } from '../../../models/GetUserByName.ts'
 import { http } from 'msw'
 
-export function getUserByNameHandler(data?: GetUserByNameQueryResponse) {
+export function getUserByNameHandler(data?: GetUserByNameQueryResponse | ((info: Parameters<Parameters<typeof http.get>[1]>[0]) => Response)) {
   return http.get('*/user/:username', function handler(info) {
+    if (typeof data === 'function') return data(info)
     return new Response(JSON.stringify(data), {
       headers: {
         'Content-Type': 'application/json',