+++ packages/plugin-msw/src/components/MockWithFaker.tsx
@@ -17,7 +17,9 @@ type Props = {
 export function MockWithFaker({ name, fakerName, typeName, url, method }: Props): ReactNode {
   const params = FunctionParams.factory({
     data: {
-      type: typeName,
+      type: `${typeName} | ((
+        info: Parameters<Parameters<typeof http.${method}>[1]>[0],
+      ) => Response)`,
       optional: true,
     },
   })
@@ -26,7 +28,9 @@ export function MockWithFaker({ name, fakerName, typeName, url, method }: Props)
     <File.Source name={name} isIndexable isExportable>
       <Function name={name} export params={params.toConstructor()}>
         {`return http.${method}('*${url}', function handler(info) {
-    return new Response(JSON.stringify(${fakerName}(data)), {
+    if(typeof data === 'function') return data(info)
+
+    return new Response(JSON.stringify(data || ${fakerName}(data)), {
       headers: {
         'Content-Type': 'application/json',
       },