+++ packages/plugin-msw/src/generators/__snapshots__/getPets.ts
@@ -1,7 +1,9 @@
 import { http } from "msw";
 
- export function listPets(data?: ListPetsQueryResponse) {
+ export function listPets(data?: ListPetsQueryResponse | ((info: Parameters<Parameters<typeof http.get>[1]>[0]) => Response)) {
     return http.get("*/pets", function handler(info) {
+        if (typeof data === "function")
+            return data(info);
         return new Response(JSON.stringify(data), {
             headers: {
                 "Content-Type": "application/json",