+++ packages/plugin-msw/src/generators/__snapshots__/getPetsFaker.ts
@@ -1,8 +1,10 @@
 import { http } from "msw";
 
- export function listPets(data?: ListPetsQueryResponse) {
+ export function listPets(data?: ListPetsQueryResponse | ((info: Parameters<Parameters<typeof http.get>[1]>[0]) => Response)) {
     return http.get("*/pets", function handler(info) {
-        return new Response(JSON.stringify(listPetsQueryResponse(data)), {
+        if (typeof data === "function")
+            return data(info);
+        return new Response(JSON.stringify(data || listPetsQueryResponse(data)), {
             headers: {
                 "Content-Type": "application/json",
             },